+++ src/components/reviews/Reviews.tsx
@@ -1,20 +1,23 @@
-import classes from './Reviwe.module.css'
+import { ContentDTO } from '../../../types/dto'
+import classes from './Content.module.css'
+import { Link } from 'react-router-dom'
 
-const Reviews = () => {
+interface IContentProps {
+  content: ContentDTO
+}
+
+const Content = ({ content }: IContentProps) => {
   return (
-    <section className={classes.container}>
-      <div className={classes.sectioncontainer}>
-        <div>
-          <div className={classes.sectionTitle}>
-            <svg xmlns="http://www.w3.org/2000/svg" height="35" viewBox="0 -960 960 960" width="35">
-              <path d="m363-390 117-71 117 71-31-133 104-90-137-11-53-126-53 126-137 11 104 90-31 133ZM80-80v-720q0-33 23.5-56.5T160-880h640q33 0 56.5 23.5T880-800v480q0 33-23.5 56.5T800-240H240L80-80Zm126-240h594v-480H160v525l46-45Zm-46 0v-480 480Z" />
-            </svg>
-            <h1 className={classes.sectionTitle}>Stories from real people</h1>
-          </div>
-        </div>
-      </div>
-      <div className={classes.blogContainer}></div>
-    </section>
+    <div className={classes.post}>
+      <Link to={`/post/${content.id}`} style={{ textDecoration: 'none', color: 'white' }}>
+        <img className={classes.thumbnailurl} src={content.thumbnailUrl}></img>
+        <p>title: {content.videoTitle}</p>
+        <h4>{content.comment}</h4>
+        <p>{content.postedBy.username}</p>
+        <p>{content.rating}</p>
+      </Link>
+    </div>
   )
 }
-export default Reviews
+
+export default Content
