--- src/components/reviews/Reviwe.module.css
@@ -1,59 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,100;0,300;0,400;0,500;0,600;1,100;1,200&family=Poppins:ital,wght@0,100;0,300;0,400;0,500;0,600;0,700;1,300&display=swap');
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: 'Kanit', sans-serif;
-}
-@media screen and (min-width: 768px) {
-  body {
-    font-size: 16px;
-  }
-}
-
-.container {
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-.blogContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 1180px;
-  height: 500px;
-  margin: auto;
-}
-
-.sectioncontainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 1180px;
-  height: 500px;
-  margin: auto;
-}
-
-.sectionTitle {
-  display: flex;
-  gap: 4px;
-  align-items: center;
-  font-size: 30px;
-  color: #343434;
-  font-weight: 100;
-  line-height: 37.5px; /* 75% */
-}
-
-.sectioncontent {
-  padding-left: 40px;
-  padding-top: 20px;
-  display: flex;
-  gap: 4px;
-  align-items: center;
-  font-size: 20px;
-  color: #4d4d4dd5;
-  font-weight: 100;
-}
